refractor: move control panel components into their own subfolder
