+++ src/components/React/control-panel/ColorMappingController.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react'
 import ReactModal from 'react-modal';
 import {connect} from 'react-redux';
-import {myChangeColorMapAction, mySaveColorMappingState} from '../../redux/AppActions'
+import {myChangeColorMapAction, mySaveColorMappingState} from '../../../redux/AppActions'
 import {DataTable} from 'primereact/datatable'
 import {Column} from 'primereact/column';
 import 'primereact/resources/themes/nova/theme.css';
